summaryLevels

Collapses factor levels from multiple columns into one and creates summary table.

Usage

summaryLevels(
  data,
  vars = NULL,
  group = NULL,
  label = NULL,
  levels = NULL,
  stat_cat = "n_percent",
  test = FALSE,
  test_cat = "fisher.test",
  ci = FALSE,
  ci_cat = "wilson",
  conf_level = 0.95,
  overall = FALSE,
  as_flex_table = TRUE
)

Arguments

data

A data frame or tibble containing the data to be summarized.

vars

Variables to include in the summary table. Default to all variables present in the data except group.

group

A single column from data. Summary statistics will be stratified according to this variable. Default to NULL.

label

A list containing the labels that should be used for the variables in the table. If NULL, labels are automaticall taken from the dataset. If no label present, the variable name is taken.

levels

= A vector containing the values indicating presence of the factor level. Included by defaul are "1", "yes", "Yes".

stat_cat

Summary statistic to display for categorical variables. Options include "n_percent" (default) and "n", and "n_N".

test

Logical. Indicates whether p-values are displayed (TRUE) or not (FALSE). Default to FALSE

test_cat

Test type used to calculated the p-value for categorical variables. Only used if test = TRUE. Options include "chisq.test", "chisq.test.no.correct", "fisher.test" (default).

ci

Logical. Indicates whether CI are displayed (TRUE) or not (FALSE). Default to FALSE.

ci_cat

Confidence interval method for categorical variables. Options include "wilson" (default), "wilson.no.correct", "clopper.pearson", "wald", "wald.no.correct", "agresti.coull" and "jeffreys". If NULL, no CI will be displayed.

conf_level

Confidence level. Default to 0.95.

overall

Logical. If TRUE, an additional column with the total is added to the table. Default to FALSE.

as_flex_table

Logical. If TRUE (default) the gtsummary object is converted to a flextable object. Useful when rendering to Word.