Creates publication-ready summary tables

Creates publication-ready summary tables based on the gtsummary package.

Usage

summaryTable(
  data,
  vars = NULL,
  group = NULL,
  labels = NULL,
  stat_cont = "median_range",
  stat_cat = "n_percent",
  continuous_as = "continuous",
  dichotomous_as = "dichotomous",
  type = NULL,
  value = NULL,
  test = FALSE,
  test_cont = "wilcox.test",
  test_cat = "fisher.test",
  ci = FALSE,
  ci_cont = "wilcox.test",
  ci_cat = "wilson",
  conf_level = 0.95,
  digits_cont = 1,
  digits_cat = 0,
  missing = TRUE,
  missing_percent = TRUE,
  missing_text = "Missing",
  overall = FALSE,
  add_n = FALSE,
  as_flex_table = TRUE
)

Arguments

data

A data frame or tibble containing the data to be summarized.

vars

Variables to include in the summary table. Default to all variables present in the data except group.

group

A single column from data. Summary statistics will be stratified according to this variable. Default to NULL.

labels

A list containing the labels that should be used for the variables in the table. If NULL, labels are automatically taken from the dataset. If no label present, the variable name is taken.

stat_cont

Summary statistic to display for continuous variables. Options include "median_IQR", "median_range" (default), "mean_sd", "mean_se" and "geomMean_sd".

stat_cat

Summary statistic to display for categorical variables. Options include "n_percent" (default) and "n", and "n_N".

continuous_as

Type for the continuous variables. Can either be "continous" (default) or "categorical".

dichotomous_as

Type for the dichotomous variables. Can either be "categorical" (default, one row per level) or "dichotomous" (only one row with reference level (see argument value), only works if missing = "FALSE" or missing_percent = FALSE.

type

NOT IMPLEMENTED YET. Specifies the summary type. Accepted values are c("continuous", "continuous2", "categorical", "dichotomous"). If not specified, default type is assigned. See below for details.

value

Specifies the reference level of a variable to display on a single row. Default is NULL. The syntax is as follows: value = list(varname ~ "level to show").

test

Logical. Indicates whether p-values are displayed (TRUE) or not (FALSE). Default to FALSE

test_cont

Test type used to calculate the p-value for continuous variables. Only used if test = TRUE. Options include "t.test", "oneway.test", "kruskal.test", "wilcox.test" (default), "paired.t.test", "paired.wilcox.test"

test_cat

Test type used to calculated the p-value for categorical variables. Only used if test = TRUE. Options include "fisher.test" (default), "chisq.test", "chisq.test.no.correct". If NULL, the function decides itself: "chisq.test.no.correct" for categorical variables with all expected cell counts >=5, and "fisher.test" for categorical variables with any expected cell count <5

ci

Logical. Indicates whether CI are displayed (TRUE) or not (FALSE). Default to FALSE.

ci_cont

Confidence interval method for continuous variables. Only used if ci = TRUE. Options include "t.test" and "wilcox.test" (default).

ci_cat

Confidence interval method for categorical variables. Options include "wilson" (default), "wilson.no.correct", "clopper.pearson", "wald", "wald.no.correct", "agresti.coull" and "jeffreys". If NULL, no CI will be displayed.

conf_level

Confidence level. Default to 0.95.

digits_cont

Digits for summary statistics and CI of continuous variables. Default to 1.

digits_cat

Digits for summary statistics and CI of categorical variables. Default to 0.

missing

Logical. If TRUE (default), the missing values are shown.

missing_percent

Indicates whether percentages for missings are shown (TRUE, default) or not (FALSE) for categorical variables. If "both", then both options are displayed next to each other.

missing_text

String indicating text shown on missing row. Default to "Missing".

overall

Logical. If TRUE, an additional column with the total is added to the table. Default to FALSE.

add_n

Logical. If TRUE, an additional column with the total number of non-missing observations for each variable.

as_flex_table

Logical. If TRUE (default) the gtsummary object is converted to a flextable object. Useful when rendering to Word.

Details

Type and values arguments By default, numeric variables with more than 2 unique values are considered continuous. This can be changed using the type argument type = list(vars ~ "categorical")

Examples

library(survival)
data("cancer")
summaryTable(data = cancer,vars = c("inst", "time","age", "ph.ecog"),
             labels = list(inst = "Institution code",
                           time = "Time",
                           age = "Age",
                           ph.ecog = "ECOG score"))

Characteristic

N = 2281

Institution code

11.0 (1.0, 33.0)

Missing

1

Time

255.5 (5.0, 1,022.0)

Age

63.0 (39.0, 82.0)

ECOG score

1.0 (0.0, 3.0)

Missing

1

1Median (Min, Max)